# openapi/openapi.json
{
  "components": {
    "schemas": {
      "CapabilitiesResponse": {
        "properties": {
          "contracts": {
            "items": {
              "$ref": "#/components/schemas/ContractCapability"
            },
            "title": "Contracts",
            "type": "array"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentCapability"
            },
            "title": "Documents",
            "type": "array"
          },
          "engines": {
            "items": {
              "$ref": "#/components/schemas/EngineCapability"
            },
            "title": "Engines",
            "type": "array"
          },
          "limits": {
            "$ref": "#/components/schemas/LimitsCapability"
          },
          "low_confidence_threshold": {
            "title": "Low Confidence Threshold",
            "type": "number"
          },
          "qr": {
            "$ref": "#/components/schemas/QRCapability"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "service",
          "version",
          "engines",
          "documents",
          "contracts",
          "limits",
          "qr",
          "low_confidence_threshold"
        ],
        "title": "CapabilitiesResponse",
        "type": "object"
      },
...
      "DIExtractRequest": {
        "additionalProperties": false,
        "properties": {
          "country": {
            "pattern": "^[A-Z]{2,3}$",
            "title": "Country",
            "type": "string"
          },
          "document_type": {
            "title": "Document Type",
            "type": "string"
          },
          "expected_response_schema_ref": {
            "const": "codestra.ocr.extraction-result/v1",
            "title": "Expected Response Schema Ref",
            "type": "string"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/DIImage"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Images",
            "type": "array"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "schema_ref": {
            "const": "codestra.ocr.extract-request/v1",
            "title": "Schema Ref",
            "type": "string"
          },
          "tenant_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.\\-]{0,63}$",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "schema_ref",
          "request_id",
          "tenant_id",
          "document_type",
          "country",
          "expected_response_schema_ref",
          "images"
        ],
        "title": "DIExtractRequest",
        "type": "object"
      },
      "DIExtractionResult": {
        "additionalProperties": false,
        "properties": {
          "country": {
            "title": "Country",
            "type": "string"
          },
          "document_type": {
            "title": "Document Type",
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DIField"
            },
            "title": "Fields",
            "type": "object"
          },
          "image_digests": {
            "$ref": "#/components/schemas/DIImageDigests"
          },
          "qr": {
            "$ref": "#/components/schemas/DIQR"
          },
          "quality": {
            "$ref": "#/components/schemas/DIQuality"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "schema_ref": {
            "const": "codestra.ocr.extraction-result/v1",
            "title": "Schema Ref",
            "type": "string"
          },
          "worker": {
            "$ref": "#/components/schemas/DIWorker"
          }
        },
        "required": [
          "schema_ref",
          "request_id",
          "worker",
          "document_type",
          "country",
          "fields",
          "image_digests",
          "quality",
          "qr"
        ],
        "title": "DIExtractionResult",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ExtractRequest": {
        "additionalProperties": false,
        "properties": {
          "country": {
            "examples": [
              "DO"
            ],
            "pattern": "^[A-Z]{2}$",
            "title": "Country",
            "type": "string"
          },
          "document_type": {
            "description": "'driver_licence' is accepted as an alias of 'driver_license'.",
            "examples": [
              "driver_license"
            ],
            "pattern": "^[a-z][a-z0-9_]{1,63}$",
            "title": "Document Type",
            "type": "string"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageInput"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Images",
            "type": "array"
          },
          "options": {
            "$ref": "#/components/schemas/ExtractOptions"
          },
          "request_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:\\-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "scan_id": {
            "description": "Caller correlation ID, echoed into the result.",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:\\-]{0,127}$",
            "title": "Scan Id",
            "type": "string"
          },
          "schema_version": {
            "examples": [
              "1.0.0"
            ],
            "pattern": "^\\d+\\.\\d+\\.\\d+$",
            "title": "Schema Version",
            "type": "string"
          },
          "tenant_id": {
            "description": "Caller tenant ID, echoed; the worker stores nothing.",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:\\-]{0,127}$",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "scan_id",
          "tenant_id",
          "document_type",
          "country",
          "schema_version",
          "images"
        ],
        "title": "ExtractRequest",
        "type": "object"
      },
      "ExtractResponse": {
        "properties": {
          "diagnostics": {
            "$ref": "#/components/schemas/Diagnostics"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/DoDriverLicenceResult"
          },
          "retention": {
            "const": "transient",
            "default": "transient",
            "title": "Retention",
            "type": "string"
          },
          "review_required": {
            "title": "Review Required",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "succeeded",
              "partial"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "status",
          "review_required",
          "result",
          "diagnostics"
        ],
        "title": "ExtractResponse",
        "type": "object"
      },
...
      "ImageInput": {
        "additionalProperties": false,
        "properties": {
          "content_base64": {
            "description": "Base64 image bytes (data: URLs accepted). Never persisted.",
            "maxLength": 22369624,
            "minLength": 16,
            "title": "Content Base64",
            "type": "string"
          },
          "media_type": {
            "anyOf": [
              {
                "enum": [
                  "image/jpeg",
                  "image/png",
                  "image/webp"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional; must match the sniffed content type if given.",
            "title": "Media Type"
          },
          "sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional lowercase hex digest; verified against the bytes.",
            "title": "Sha256"
          },
          "side": {
            "default": "unknown",
            "description": "Card side. With two 'unknown' images the first is treated as front.",
            "enum": [
              "front",
              "back",
              "unknown"
            ],
            "title": "Side",
            "type": "string"
          }
        },
        "required": [
          "content_base64"
        ],
        "title": "ImageInput",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Internal-only service. Accepts bounded document images, runs a pluggable OCR engine (Tesseract + OpenCV preprocessing by default) and returns transient, schema-mapped extraction results with per-field confidence and evidence. It never persists images, never fetches QR URLs and does not own client or business records.",
    "summary": "Internal OCR worker runtime and engine adapter layer.",
    "title": "Codestra OCR Worker",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/ocr/extract": {
      "post": {
        "operationId": "extract_internal_v1_ocr_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid internal token"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Payload or image exceeds limits"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, image or unsupported document/schema"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "OCR engine failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Worker busy or engine unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Extraction deadline exceeded"
          }
        },
        "summary": "Extract fields; result conforms to Codestra-Document-Schemas v1",
        "tags": [
          "ocr"
        ]
      }
    },
...
    "/v1/capabilities": {
      "get": {
        "operationId": "capabilities_v1_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid internal token"
          }
        },
        "summary": "Engines, document types, schema versions, contracts and limits",
        "tags": [
          "capabilities"
        ]
      }
    },
    "/v1/ocr/extract": {
      "post": {
        "operationId": "extract_di_v1_ocr_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DIExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DIExtractionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid internal token"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Payload or image exceeds limits"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, image or unsupported document/schema"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "OCR engine failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Worker busy or engine unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Extraction deadline exceeded"
          }
        },
        "summary": "Document-Intelligence adapter (codestra.ocr.extract-request/v1)",
        "tags": [
          "ocr"
        ]
      }
    }
  }
}